From Andreas Morsch
thanks Marcello Lunardon for reporting this. I found the problem. In the interface to
G3 the component of the information stored in the g3-momentum vector
which is the total momentum has been interpreted as energy and passed
as such to the decayer. The decayer itself is ok.
If e < m, instead of complaining, Pythia replaces the energy by the mass
hiding part of the problem.