Add a description of the gcalor interface ========================================= The GCALOR package contains the original CALOR simulation code (including the low energy neutron code MICAP ) developed and maintained by the Oak Ridge National Laboratory ( ORNL ) and an interface to utilize this code in the framework of GEANT . GCALOR consists of
HETC
The High-Energy-Transport-Code is transporting charged hadrons up to an energy of 10 GeV through the materials of the setup
FLUKA
The FLUKA fragmentation model is utilized for interactions above the HETC limit
MICAP
The neutron code from the Monte-carlo-Ionization-Chamber-Analysis-Program is called for neutrons with a kinetic energy below 20 MeV. The simulation of interactions is based on measured/calculated cross sections and secondary particle energy/angular distributions

Interface
---------
The interface between CALOR and GEANT determines which code to call for the current particle. The material description is automatically extracted from GEANT The transport of electromagnetic interacting particles is performed by GEANT as well as the tracking and the calculation of the energy deposition.

Installation:
------------
 1. download the CALOR code (see below) and copy it to  geant3/gcalor directory
 2. in the geant3/Makefile  add the gcalor directory to the GDIRS
recompile geant3

Usage with vmc
--------------
 one has call TGeant3::SetHADR(5) in the initalization of Geant3.


For the initialisation, GCALOR needs the following  additional cross-section data (expected files: chetc.dat and xsneut.dat in the $G3SYS/lib directory) for all needed materials, GCALOR Calculate the macroscopic cross-section for low energy neutrons and write relevant informations for the user to file calor.out



The GCALOR package has been tested on different computer platforms:

Unix systems
DecAlpha, Silicon Graphics, Hewlett Packard, IBM-R6000, SUN Sparc and Linux on PC
DecVMS
DecALpha and VAX
MVS
IBM Mainframe 3090

The GCALOR does not work for the moment on 64 bit arcitecture. For the 64 bit, it is needed to compile with the option "-m32".


For download and  more detailed information:  http://www.staff.uni-mainz.de/zeitnitz/Gcalor/gcalor.html