From Lia.Lavezzi@pv.infn.it  and alberto.rotondi@pv.infn.it
in PANDA we prepared some additional bug fixes for GEANE

There are two corrections:
1) the propagation to length used to contain a bug so that if the
propagation to length was chosen but the given length was too small
(or 0, as limit case) it did not work.
We fixed it in such a way that if the track length is too small but not 0
the step size is set to PREC, if it is <= 0 it is set to 0 and the start
point is coincident with the end point.
The flag IERR has been provided of the additional value 5: in this case,
when start and end position coincide, IERR is set to 5 and this means that
the error matrix does not have to be updated. In this way if track length
is 0 the particle just does not move.
2) it has been added the possibility to track along the z axis, though
with an approximation. Up to now, due to the matrix elements with
1/cos(lambda), the propagation exactly along z axis (or below a certain
limit) was forbidden. The approximation we introduced is in TGeant3gu.gxx,
in eustep() function: if the theta angle drops below 1e-4 deg it is forced
to this limit value in order to be able to go on tracking. Obviously this
means that if you try to track exactly along z (theta = 0) you will be
forced to track at 1e-4 deg instead. 

We checked the pulls and residual values for a plane @12m and @26m and 
they both with and without magnetic field and they are acceptable.