Detect if primary when popping
With this commit G3 detects whether a particle is a primary depending on
the return value TParticle::IsPrimary()
The information is propagated to gtreveroot.F and used for the decision
whether RXOUTH or RXINH should be called.
Hence, the commit makes this decision independent on the user numbering
scheme of stacked particles.