Make ready for multiple engine usage
With this commit GEANT3 becomes capable of running together with other
engines sharing the simulation of one event. Depending on user
conditions, tracks might be transported by an engine based on geometry
constraints, particle type of phase space. When these conditions are not
met anymore, the VMC TMCManager interrupts the transport and transfers
the track to the engine responsible under the new conditions. This can
happen multiple times and is done automatically until all tracks have
been fully transported.