Revert G3 internal track indexing
Commit e1a8a4424bab11cee668d094a29931d1d542841d breaks the single engine
run due to the assumptions made one the G3 internal track indexing and
the order in which tracks are transported. Hence, this has to be
reverted.
Adjustments for multi engine scenario are necessary on VMC level
strictly complying with the current stacking logic and are on the way.

The single engine scenario works now as usual.