Fix compiler warnings
Most warnings were related to array size mismatches.
In cases an array is passed to a subroutine don't define a fixed array size
in the subroutine but let the compiler figure out the correct size.
One warning(error?) was due to a typo in the variable name.
Some warnings were due to an agressive loop optimization. Switch off this
xcompiler feature for the compilation of gcalor.