Fix linking on macosx12
The problems does not appear during the actual linking but at runtime when
loading the library. In this case the dynamic linker isn"t able to find the
needed pythia6 library to resolve all symbols. This problem is fixed by adding
a new dependency to libEGpythia6 which itself depends on pythia6. This
indirect dependency avoids adding a macro which directly searches for pyhia6.