Patch mutiple volumes to sd (#34)
The ROOT Geometry package does not have any restriction on volumes sharing the same name.

When there are multiple sensitive volumes with the same name, sensitive detectors should be registered to all of them.

This PR replaces the single volume lookup to a loop over all volumes.

It also elevates the severity of the case where a user's volume extension is replaced. This now indicates an Error condition, rather than just a warning. The ROOT/TGeoVolume documentation strongly implies that the "user extension" mechanism is available for the exclusive use of the user, not the framework. When a volume extension is overwritten, it is likely to cause problems in the rest of the user's code.